Refactoring: job-API related tests
